+++ src/__tests__/integration.js
@@ -6,6 +6,7 @@ const config = {
   rules: {
     "block-opening-brace-newline-after": [ 2, "always" ],
     "color-no-invalid-hex": 2,
+    "selector-pseudo-element-colon-notation": [ 2, "double" ],
   },
 }
 
@@ -22,7 +23,7 @@ b { background: pink; }
 }
 /* stylelint-enable */
 
-.bar {
+.bar:before {
   color: #mmm;
 }
 `)
@@ -36,12 +37,13 @@ postcss()
 function checkResult(result) {
   const { messages } = result
   test("expected warnings", t => {
-    t.equal(messages.length, 3)
+    t.equal(messages.length, 4, "number of errors")
     t.ok(messages.every(m => m.type === "warning"))
     t.ok(messages.every(m => m.plugin === "stylelint"))
-    t.equal(messages[0].text, "Expected newline after \"{\" (block-opening-brace-newline-after)")
-    t.equal(messages[1].text, "Unexpected invalid hex color \"#zzz\" (color-no-invalid-hex)")
-    t.equal(messages[2].text, "Unexpected invalid hex color \"#mmm\" (color-no-invalid-hex)")
+    t.equal(messages[0].text, "Expected newline after \"{\" (block-opening-brace-newline-after)", "first error message")
+    t.equal(messages[1].text, "Unexpected invalid hex color \"#zzz\" (color-no-invalid-hex)", "second error message")
+    t.equal(messages[2].text, "Unexpected invalid hex color \"#mmm\" (color-no-invalid-hex)", "fourth error message")
+    t.equal(messages[3].text, "Expected double colon pseudo-element notation (selector-pseudo-element-colon-notation)", "third error message")
     t.end()
   })
 }