+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Head
+
+* Added: error about undefined severities blaming stylelint for the bug.
+* Fixed: `selector-pseudo-element-colon-notation` typo in rule name resulting in undefined severity.
+
 # 2.3.4
 
 * Fixed: `dist/` build.